+++ src/modules/elements.js
@@ -1,0 +1,8 @@
+const todoList = document.querySelector('.todo-list');
+const form = document.querySelector('.form');
+const taskInput = document.querySelector('.task-input');
+const clearButton = document.querySelector('.clear-button');
+
+export {
+  todoList, form, taskInput, clearButton,
+};
