+++ src/modules/storage.js
@@ -1,0 +1,6 @@
+const SAVE_LOC = 'task.lists';
+const tasks = JSON.parse(localStorage.getItem(SAVE_LOC)) || [];
+const save = () => {
+  localStorage.setItem(SAVE_LOC, JSON.stringify(tasks));
+};
+export { tasks, save };
