+++ src/style.css
@@ -27,7 +27,8 @@ ul {
 
 li,
 .task-input,
-.list-header {
+.list-header,
+.edit-input {
   display: flex;
   justify-content: space-between;
   align-items: center;
@@ -37,7 +38,7 @@ li,
 }
 
 label {
-  padding: 0.8rem;
+  padding: 0 0.8rem;
 }
 
 div {
@@ -114,6 +115,17 @@ button {
   width: 100%;
 }
 
+.edit-input {
+  border: none;
+  outline: none;
+  font-size: 1rem;
+  width: 100%;
+  padding: 0;
+  margin: 0;
+  margin-left: 0.8rem;
+  background-color: inherit;
+}
+
 .clear-button {
   display: flex;
   justify-content: center;
@@ -132,6 +144,12 @@ button {
   background-color: inherit;
 }
 
+.more-icon.del-btn {
+  margin: 0;
+  padding: 0;
+  padding: 0 0.28rem;
+}
+
 .task-desc {
   cursor: pointer;
 }
@@ -142,8 +160,8 @@ button {
 
 .disappear {
   display: none;
-  
 }
+
 .appear {
-  display: '';
+  display: block;
 }
